+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 - Changed [README.md][] Translate Japanese to English.
+- Added [Get-Dataset.py][] function.
 
 ## [0.1.0] - 2025-02-21
 
@@ -34,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Calc-ChemMassPercent.py]: src/Calc-ChemMassPercent.py
 
 [Calc-LPpulp.py]: src/Calc-LPpulp.py
+
+[Get-Dataset.py]: src/Get-Dataset.py
 
 
 [unreleased]: https://github.com/btklab/python-sketches/compare/0.1.0..HEAD